add ModalityCheckBox and TriStateModalityCheckBox components (previously part of DatasetSamplesTable) for reuse in ProjectSamplesTable
